Rename files & amend includes
